requirements.txt
sentence-transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain_community
langchain_openai
langchain_experimental
langchain_pinecone
setuptools

-e .